+++ src/nametables/enter_high_score.js
@@ -31,8 +31,6 @@ ui!###g@######()
 
 lookup = [...lookup].map((d, i) => d === '#' ? String.fromCharCode(9472 + i) : d).join``;
 
-printNT(buffer, lookup);
-
 drawTiles(buffer, lookup, `
 ################################
 #qwwwwwwwwwwwwwwwwwwwwwwwwwwwwe#